# client/schemas/openapi.json
{"openapi":"3.0.0","paths":{"/todo":{"post":{"operationId":"TodoController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTodoDto"}}}},"responses":{"200":{"description":"Greate a new todo","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Todo"}}}},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DefaultException"}}}}},"tags":["todo"],"security":[{"cookie":[]}]}},"/todo/query":{"get":{"operationId":"TodoController_query","parameters":[{"name":"status","required":false,"in":"query","schema":{"enum":["notStarted","inProgress","completed"],"type":"string"}},{"name":"dueDate","required":false,"in":"query","schema":{"format":"date-time","type":"string"}},{"name":"sortBy","required":false,"in":"query","schema":{"type":"string"}},{"name":"sortOrder","required":false,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"Query todos","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Todo"}}}}},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DefaultException"}}}}},"tags":["todo"],"security":[{"cookie":[]}]}},"/todo/{id}":{"get":{"operationId":"TodoController_findOneById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"get a todo by id","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Todo"}}}},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DefaultException"}}}}},"tags":["todo"],"security":[{"cookie":[]}]},"patch":{"operationId":"TodoController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateTodoDto"}}}},"responses":{"200":{"description":"update a todo by id","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Todo"}}}},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DefaultException"}}}}},"tags":["todo"],"security":[{"cookie":[]}]},"delete":{"operationId":"TodoController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"201":{"description":"delete a todo by id"},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DefaultException"}}}}},"tags":["todo"],"security":[{"cookie":[]}]}},"/auth/register":{"post":{"operationId":"AuthController_register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterDto"}}}},"responses":{"200":{"description":"register","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DefaultException"}}}}},"tags":["auth"]}},"/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DefaultException"}}}},"default":{"description":"Login","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginResponseDto"}}}}},"tags":["auth"]}},"/auth/logout":{"post":{"operationId":"AuthController_logout","parameters":[],"responses":{"201":{"description":""}},"tags":["auth"]}}},"info":{"title":"Api for TODO management system","description":"The API description","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"cookie":{"type":"apiKey","in":"cookie","name":"access_token"}},"schemas":{"TodoStatus":{"type":"string","enum":["notStarted","inProgress","completed"]},"CreateTodoDto":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"},"dueDate":{"format":"date-time","type":"string","nullable":true},"status":{"$ref":"#/components/schemas/TodoStatus"}},"required":["name"]},"Todo":{"type":"object","properties":{"id":{"type":"number"},"userId":{"type":"number"},"name":{"type":"string"},"description":{"type":"string"},"dueDate":{"format":"date-time","type":"string","nullable":true},"status":{"$ref":"#/components/schemas/TodoStatus"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"},"deletedAt":{"format":"date-time","type":"string","nullable":true}},"required":["id","userId","name","description","dueDate","status","createdAt","updatedAt","deletedAt"]},"DefaultException":{"type":"object","properties":{"statusCode":{"type":"number"},"message":{"oneOf":[{"type":"string"},{"type":"array","items":{"type":"string"}}]},"error":{"type":"string"}},"required":["statusCode","message","error"]},"UpdateTodoDto":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"},"dueDate":{"format":"date-time","type":"string","nullable":true},"status":{"$ref":"#/components/schemas/TodoStatus"}}},"RegisterDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]},"User":{"type":"object","properties":{"id":{"type":"number"},"email":{"type":"string"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"},"deletedAt":{"format":"date-time","type":"string"}},"required":["id","email","createdAt","updatedAt","deletedAt"]},"LoginDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]},"LoginResponseDto":{"type":"object","properties":{"accessToken":{"type":"string"}},"required":["accessToken"]}}}}